requirements.txt
alembic==1.16.5
altair==4.2.2
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
beautifulsoup4==4.14.2
bleach==6.2.0
blinker==1.9.0
boto3==1.40.49
botocore==1.40.49
cachetools==6.2.0
category_encoders==2.8.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
codecarbon==3.0.5
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
cyclopts==3.24.0
dacite==1.6.0
dagshub==0.6.3
dagshub-annotation-converter==0.1.13
databricks-sdk==0.67.0
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
deepchecks==0.19.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.118.2
fastjsonschema==2.21.2
fastmcp==2.12.4
fief-client==0.20.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.0
fqdn==1.5.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
gql==4.0.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
Gx==1.0.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.2
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
jwcrypto==1.5.6
kiwisolver==1.4.9
lark==1.3.0
lazy-object-proxy==1.12.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
mistune==3.1.4
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-ml-py==13.580.82
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pathvalidate==3.3.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
polars==1.33.1
polars-lts-cpu==1.33.1
posthog==6.7.6
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyNomaly==0.3.4
pynvml==13.0.1
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.20
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
questionary==2.1.1
RapidFuzz==3.14.1
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
semver==3.0.4
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toolz==1.0.0
torch==2.2.2
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treelib==1.8.0
typer==0.19.2
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.203
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.1
widgetsnbextension==4.0.14
yarl==1.22.0
yaspin==3.2.0
zipp==3.23.0
